refactor(rust): split visibility and path reasoning out of the resolver
The resolver answered three questions in one file: which declarations a usage can
see, what role a name plays in a `::` path, and which candidate it names. The first
two are self-contained and now say so.

  visibility        reachable, and which of them is nearest
  qualified_paths   which segment of a path a name is
  rust_dependency_resolver   what a usage resolves to

Behaviour is unchanged: 743 of 743, precision and recall 1.000, identical to the
recorded baseline.
